+++ common/src/main/java/net/spell_engine/api/datagen/SpellGenerator.java
@@ -14,6 +14,7 @@ import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Util;
 import net.spell_engine.api.spell.Spell;
+import net.spell_engine.api.spell.fx.ModelEffect;
 import net.spell_engine.api.spell.fx.PlayerAnimation;
 import net.spell_engine.api.spell.fx.ParticleBatch;
 import net.spell_engine.api.spell.fx.Sound;
@@ -84,7 +85,8 @@ public abstract class SpellGenerator implements DataProvider {
                 .registerTypeAdapter(Spell.class, new DefaultValueSkippingSerializer<>(Spell.class))
                 .registerTypeAdapter(ParticleBatch.class, new DefaultValueSkippingSerializer<>(ParticleBatch.class))
                 .registerTypeAdapter(Sound.class, new DefaultValueSkippingSerializer<>(Sound.class))
-                .registerTypeAdapter(PlayerAnimation.class, new DefaultValueSkippingSerializer<>(PlayerAnimation.class));
+                .registerTypeAdapter(PlayerAnimation.class, new DefaultValueSkippingSerializer<>(PlayerAnimation.class))
+                .registerTypeAdapter(ModelEffect.class, new DefaultValueSkippingSerializer<>(ModelEffect.class));
         for (var nestedClass : getAllNestedClasses(Spell.class)) {
             gson = gson.registerTypeAdapter(nestedClass, new DefaultValueSkippingSerializer<>(nestedClass));
         }